TTNN Runtime refactor
