+++ runtime/lib/ttnn/operations/context/get_device/get_device.h
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: (c) 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef TTNN_RUNTIME_GET_DEVICE_H
+#define TTNN_RUNTIME_GET_DEVICE_H
+
+#include "tt/runtime/ttnn/types.h"
+#include "ttmlir/Target/TTNN/program_generated.h"
+
+namespace tt::runtime::ttnn::operations::context {
+void run(const ::tt::target::ttnn::GetDeviceOp *op, ProgramContext &context);
+
+} // namespace tt::runtime::ttnn::operations::context
+
+#endif
