+++ runtime/lib/ttnn/operations/data_movement/concat/concat.cpp
@@ -1,0 +1,18 @@
+// SPDX-FileCopyrightText: (c) 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "concat.h"
+#include "tt/runtime/detail/ttnn.h"
+
+namespace tt::runtime::ttnn::operations::data_movement {
+void run(const ::tt::target::ttnn::ConcatOp *op, ProgramContext &context) {
+  std::vector<::ttnn::Tensor> inputs;
+  for (const auto &input : *op->inputs()) {
+    inputs.push_back(context.tensorPool.at(input->global_id()));
+  }
+  int32_t dim = op->dim();
+  ::ttnn::Tensor out = ::ttnn::concat(inputs, dim);
+  context.tensorPool.insert_or_assign(op->out()->global_id(), out);
+}
+} // namespace tt::runtime::ttnn::operations::data_movement
