+++ runtime/lib/ttnn/operations/data_movement/transpose/transpose.cpp
@@ -1,0 +1,40 @@
+// SPDX-FileCopyrightText: (c) 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "transpose.h"
+#include "tt/runtime/detail/ttnn.h"
+#include "tt/runtime/ttnn/operations/utils.h"
+
+namespace tt::runtime::ttnn::operations::data_movement {
+void run(const ::tt::target::ttnn::TransposeOp *op, ProgramContext &context) {
+  ProgramTensorPool &tensorPool = context.tensorPool;
+  const ::ttnn::Tensor &in = tensorPool.at(op->in()->global_id());
+  int32_t dim0 = op->dim0();
+  int32_t dim1 = op->dim1();
+  auto inputRank = in.get_shape().rank();
+  // for the current version of permute, we need to work in 4D, so we add
+  // leading dimensions of size 1
+  std::vector<std::int64_t> dimensionOrder(4);
+  std::iota(dimensionOrder.begin(), dimensionOrder.end(), 0);
+  if (dim0 < 0) {
+    dim0 += 4;
+  } else {
+    dim0 = dim0 + 4 - inputRank;
+  }
+  if (dim1 < 0) {
+    dim1 += 4;
+  } else {
+    dim1 = dim1 + 4 - inputRank;
+  }
+  std::swap(dimensionOrder[dim0], dimensionOrder[dim1]);
+  // Ideally this would use ttnn::transpose, but since ttnn::transpose doesn't
+  // work at the moment, we use this temporary solution.
+  ::ttnn::Tensor unsqueezedInput = ::ttnn::unsqueeze_to_4D(in);
+  ::tt::tt_metal::MemoryConfig outputMemoryConfig =
+      utils::createMemoryConfig(op->out());
+  ::ttnn::Tensor out =
+      ::ttnn::permute(unsqueezedInput, dimensionOrder, outputMemoryConfig);
+  tensorPool.insert_or_assign(op->out()->global_id(), out);
+}
+} // namespace tt::runtime::ttnn::operations::data_movement
