+++ runtime/lib/ttnn/operations/layout/to_memory_config/to_memory_config.h
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: (c) 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef TTNN_RUNTIME_TO_MEMORY_CONFIG_H
+#define TTNN_RUNTIME_TO_MEMORY_CONFIG_H
+
+#include "tt/runtime/ttnn/types.h"
+#include "ttmlir/Target/TTNN/program_generated.h"
+
+namespace tt::runtime::ttnn::operations::layout {
+void run(const ::tt::target::ttnn::ToMemoryConfigOp *op,
+         ProgramContext &context);
+} // namespace tt::runtime::ttnn::operations::layout
+
+#endif
